requirements.txt
pdf2image==1.9.0
Pillow==6.2.0
PyPDF2==1.26.0
pytesseract==0.3.0
